requirements.txt
confluent_kafka
psutil
requests
rocketmq-client-python